COMP: Fixing dashboard compile warnings on type conversion.

+++ VTK/Infovis/vtkFast2DLayoutStrategy.cxx
@@ -472,8 +472,8 @@ void vtkFast2DLayoutStrategy::ResolveCoincidentVertices()
   // Place the vertices into a giant grid (100xNumVertices)
   // and see if you have any collisions
   vtkBitArray *giantGrid = vtkBitArray::New();
-  vtkIdType xDim = sqrt((float)numVertices) * 10;
-  vtkIdType yDim = sqrt((float)numVertices) * 10;
+  vtkIdType xDim = static_cast<int>(sqrt((float)numVertices) * 10);
+  vtkIdType yDim = static_cast<int>(sqrt((float)numVertices) * 10);
   vtkIdType gridSize = xDim * yDim;
   giantGrid->SetNumberOfValues(gridSize);
